+++ packages/cli/src/commands/machine/machineTranslateBundle.ts
@@ -1,19 +1,22 @@
 import { randomUUID } from "node:crypto";
 import {
   Text,
-  VariableReference,
   type BundleNested,
   type NewBundleNested,
   type Variant,
 } from "@inlang/sdk";
-
-const PRIMARY_API_KEY_ENV = "INLANG_GOOGLE_TRANSLATE_API_KEY";
+import {
+  deserializePattern,
+  findMatchingVariant,
+  serializePattern,
+} from "./patternSerialization.js";
+import type { MachineTranslateProvider } from "./providers/types.js";
 
 type MachineTranslateArgs = {
   bundle: BundleNested;
   sourceLocale: string;
   targetLocales: string[];
-  googleApiKey?: string;
+  provider: MachineTranslateProvider;
 };
 
 export type MachineTranslateResult = {
@@ -22,17 +25,18 @@ export type MachineTranslateResult = {
 };
 
 /**
- * Machine translates the given bundle using the Google Translate API.
+ * Machine translates the given bundle using the configured translation provider.
  *
  * The translation updates or creates variants only for missing translations in
  * the requested target locales. Existing non-empty variants are preserved.
  *
  * @example
+ *   const provider = resolveMachineTranslateProvider();
  *   const result = await machineTranslateBundle({
  *     bundle,
  *     sourceLocale: "en",
  *     targetLocales: ["de"],
- *     googleApiKey: process.env.INLANG_GOOGLE_TRANSLATE_API_KEY,
+ *     provider,
  *   });
  *   if (result.data) {
  *     await upsertBundleNested(project, result.data);
@@ -42,13 +46,6 @@ export async function machineTranslateBundle(
   args: MachineTranslateArgs,
 ): Promise<MachineTranslateResult> {
   try {
-    const apiKey =
-      args.googleApiKey ?? process.env[PRIMARY_API_KEY_ENV];
-
-    if (!apiKey) {
-      return { error: `${PRIMARY_API_KEY_ENV} is not set` };
-    }
-
     const copy = structuredClone(args.bundle);
 
     const sourceMessage = copy.messages.find(
@@ -92,27 +89,17 @@ export async function machineTranslateBundle(
           }
         }
 
-        const response = await fetch(
-          "https://translation.googleapis.com/language/translate/v2?" +
-            new URLSearchParams({
-              q: sourcePattern,
-              target: targetLocale,
-              source: args.sourceLocale,
-              format: "html",
-              key: apiKey,
-            }),
-          { method: "POST" },
-        );
+        const translation = await args.provider.translateText({
+          text: sourcePattern,
+          sourceLocale: args.sourceLocale,
+          targetLocale,
+        });
 
-        if (!response.ok) {
-          const err = `${response.status} ${response.statusText}: translating from ${args.sourceLocale} to ${targetLocale}`;
-          return { error: err };
+        if (!translation.ok) {
+          return { error: translation.error };
         }
 
-        const json = await response.json();
-        const pattern = deserializePattern(
-          json.data.translations[0].translatedText,
-        );
+        const pattern = deserializePattern(translation.translatedText);
 
         if (targetMessage) {
           const existingVariant = findMatchingVariant(
@@ -159,113 +146,4 @@ export async function machineTranslateBundle(
   } catch (error) {
     return { error: error?.toString() ?? "unknown error" };
   }
-}
-
-function findMatchingVariant(
-  variants: Variant[],
-  matches: Variant["matches"],
-): Variant | undefined {
-  if (matches.length === 0) {
-    return variants.find((variant) => variant.matches.length === 0);
-  }
-
-  return variants.find((variant) => {
-    if (variant.matches.length !== matches.length) {
-      return false;
-    }
-
-    return matches.every((sourceMatch) =>
-      variant.matches.some((targetMatch) => {
-        if (
-          targetMatch.key !== sourceMatch.key ||
-          targetMatch.type !== sourceMatch.type
-        ) {
-          return false;
-        }
-
-        if (
-          sourceMatch.type === "literal-match" &&
-          targetMatch.type === "literal-match"
-        ) {
-          return sourceMatch.value === targetMatch.value;
-        }
-
-        return true;
-      }),
-    );
-  });
-}
-
-type PlaceholderMetadata = Record<
-  string,
-  {
-    leadingCharacter?: string;
-    trailingCharacter?: string;
-  }
->;
-
-const escapeStart = `<span class="notranslate">`;
-const escapeEnd = "</span>";
-
-function serializePattern(
-  pattern: Variant["pattern"],
-  placeholderMetadata: PlaceholderMetadata,
-) {
-  let result = "";
-  for (const [index, element] of pattern.entries()) {
-    if (element.type === "text") {
-      result += element.value
-        .replaceAll("\r", "<inlang-CarriageReturn>")
-        .replaceAll("\n", "<inlang-LineFeed>");
-    } else {
-      // @ts-expect-error placeholder metadata is keyed by name at runtime
-      placeholderMetadata[element.name] = {
-        leadingCharacter: result.at(-1) ?? undefined,
-        trailingCharacter:
-          pattern[index + 1]?.type === "text"
-            ? (pattern[index + 1] as Text).value[0]
-            : undefined,
-      };
-      result += `${escapeStart}${JSON.stringify(element)}${escapeEnd}`;
-    }
-  }
-  return result;
-}
-
-function deserializePattern(text: string): Variant["pattern"] {
-  const result: Variant["pattern"] = [];
-  const unescapedText = text
-    .replaceAll("&quot;", '"')
-    .replaceAll("&#39;", "'")
-    .replaceAll("<inlang-CarriageReturn>", "\r")
-    .replaceAll("<inlang-LineFeed> ", "\n")
-    .replaceAll("<inlang-LineFeed>", "\n");
-  let index = 0;
-  while (index < unescapedText.length) {
-    const start = unescapedText.indexOf(escapeStart, index);
-    if (start === -1) {
-      result.push({ type: "text", value: unescapedText.slice(index) });
-      break;
-    } else if (index < start) {
-      result.push({ type: "text", value: unescapedText.slice(index, start) });
-      index = start;
-      continue;
-    }
-    const end = unescapedText.indexOf(escapeEnd, start);
-    if (end === -1) {
-      result.push({ type: "text", value: unescapedText.slice(index) });
-      break;
-    }
-
-    const expressionAsText = unescapedText.slice(
-      start + escapeStart.length,
-      end,
-    );
-    const expression = JSON.parse(expressionAsText) as VariableReference;
-
-    // @ts-expect-error placeholder metadata is preserved at runtime
-    result.push(expression);
-    index = end + escapeEnd.length;
-  }
-  return result;
 }
